confused-backend/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
astroid==2.12.9
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==21.4b2
bootstrap4==0.1.0
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
charset-normalizer==2.0.12
click==7.1.2
constantly==15.1.0
cryptography==37.0.1
cssselect2==0.6.0
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.0.4
django-appconf==1.0.5
django-bootstrap4==22.1
django-classy-tags==3.0.1
django-compressor==4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-formtools==2.3
django-heroku==0.3.1
django-libsass==0.9
django-model-utils==4.2.0
django-otp==1.1.3
django-sekizai==3.0.1
django-session-security==2.6.6
django-tables2==2.4.1
django-treebeard==4.5.1
django-wkhtmltopdf==3.4.0
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
flake8==3.9.2
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.10.1
lazy-object-proxy==1.7.1
libsass==0.21.0
lxml==4.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
num2words==0.5.10
oscrypto==1.3.0
pathspec==0.8.1
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydyf==0.1.2
pyflakes==2.3.1
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
PyPDF2==1.27.9
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
rcssmin==1.1.0
regex==2021.4.4
reportlab==3.6.9
requests==2.27.1
rjsmin==1.2.0
rope==0.19.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
xhtml2pdf==0.2.7
zope.interface==5.4.0
zopfli==0.2.1
